change 'type' key to 'level' in version.py filter plugin.
'type' is still there as an alias to help in transition,
with the intention to remove it after dependent roles are
switched over.

i've moved to using the term 'level' for the string after the
`M.m.p` in versions (the 'rc' in `0.1.2-rc.3`, 'dev' in `0.1.2-dev`,
etc.), and wanted to switch this over as well.